pyproject.toml
[project]
name = "hailstack"
version = "0.1.0"
description = "Provision and manage Spark/Hadoop/Hail clusters on OpenStack."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.5",
    "PyYAML>=6.0.3",
    "python-dotenv>=1.2.2",
    "typer>=0.16.0",
    "pulumi>=3.226.0",
    "pulumi-openstack>=5.4.1",
]

[project.scripts]
hailstack = "hailstack.cli.main:app"

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hailstack"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"
include = ["src"]
venvPath = "."
venv = ".venv"
